Disable FASTMATH version of sqrt for doubles
The AVX512 FASTMATH implementation of sqrt does not seem to produce
accurate enough results for some of the TF unit tests.  More investigation
is needed to see whether it can be fixed but for the time being we're
just going to disable it.  Note that there is no FASTMATH sqrt implementation
for doubles in the AVX2 code, so disabling it may well turn out to be the
correct thing to do.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>